requirements.txt
beautifulsoup4==4.8.2
dj-database-url==0.5.0
Django==3.0.5
django-heroku==0.3.1
djangorestframework==3.11.0
pandas==1.0.1
Pillow==5.1.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dateutil==2.8.1
decorator==4.4.2
distlib==0.3.0
requests==2.23.0
simplejson==3.13.2
googletrans==2.4.0
SQLAlchemy==1.1.11
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
gunicorn==20.0.4